chore(api)!: Unexport some Config methods
